requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
flexcache==0.3
flexparser==0.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
MarkupSafe==3.0.2
mathparse==0.1.2
nltk==3.9.1
openai==1.52.0
Pint==0.24.3
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==3.13.0
python-dateutil==2.7.5
python-dotenv==1.0.1
pytz==2024.2
PyYAML==3.13
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.2.19
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
